Refactored names and imports - Changed names of plugin files to match the rest - Reordered and separated imports into categories
