+++ src/plugins/ApiClient.ts
@@ -1,10 +1,10 @@
 import axios, { AxiosInstance } from 'axios';
 
-const apiClient: AxiosInstance = axios.create({
+const ApiClient: AxiosInstance = axios.create({
   baseURL: process.env.VUE_APP_BASE_API_URL,
   headers: {
     'Content-type': 'application/json'
   }
 });
 
-export default apiClient;
+export default ApiClient;
