+++ src/plugins/FontAwesomeIcons.ts
