+++ src/services/CommonDataService.ts
@@ -1,11 +1,11 @@
+import ApiClient from '@/plugins/ApiClient';
 import HomeBox from '@/models/HomeBox';
 import Navigation from '@/models/Navigation';
-import apiClient from '@/plugins/apiClient';
 
 export default {
   async getMenu(): Promise<Navigation[]> {
     try {
-      return (await apiClient.get('/menu')).data;
+      return (await ApiClient.get('/menu')).data;
     } catch (error) {
       console.error(error);
       return [];
@@ -14,7 +14,7 @@ export default {
 
   async getHomeBoxes(): Promise<HomeBox[]> {
     try {
-      return (await apiClient.get('/homeBoxes')).data;
+      return (await ApiClient.get('/homeBoxes')).data;
     } catch (error) {
       console.error(error);
       return [];
