+++ src/views/Campaigns.vue
@@ -4,6 +4,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
+
 import CommonTitle from '@/components/CommonTitle.vue';
 
 export default defineComponent({
