+++ src/views/Cultures.vue
@@ -4,7 +4,9 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
+
 import Table from '@/models/Table';
+
 import TableWithRouter from '@/components/TableWithRouter.vue';
 
 export default defineComponent({
