+++ src/views/Locations.vue
@@ -4,8 +4,10 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import TableWithRouter from '@/components/TableWithRouter.vue';
+
 import Table from '@/models/Table';
+
+import TableWithRouter from '@/components/TableWithRouter.vue';
 
 export default defineComponent({
   name: 'Locations',
