+++ src/views/Releases.vue
@@ -4,8 +4,10 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import TableWithRouter from '@/components/CommonTitle.vue';
+
 import Table from '@/models/Table';
+
+import TableWithRouter from '@/components/CommonTitle.vue';
 
 export default defineComponent({
   name: 'Releases',
